refactor(css/parser): code
